Application에 있던 자료들 Lotto 클래스로 옮김
